ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
chromadb
psycopg2-binary
beautifulsoup4
requests
grpcio
grpcio-tools
huggingface-hub
